app/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
passlib==1.7.1
Pillow==5.4.0
pkg-resources==0.0.0
Werkzeug==0.14.1
